Implemented score
Added score, which increases with each enemy killed. Bonuses are also
awarded for beating the level, how much money is left over, and how much
health the castle has left at the end.

+++ Assets/Scripts/MoneyManager.cs
@@ -5,7 +5,7 @@ using UnityEngine.UI;
 
 public class MoneyManager : MonoBehaviour
 {
-    public float balance = 100;
+    public int balance = 100;
     public GameObject towerPrefab;
     Text balanceTxt;
     public bool placeMode = false;
